+++ src/main/java/calculator/view/OutputView.java
@@ -4,7 +4,7 @@ public class OutputView {
 
     private static final String RESULT_MESSAGE = "결과 : ";
 
-    public void printResult(int result) {
+    public void printResult(long result) {
         System.out.println(RESULT_MESSAGE + result);
     }
 }
